Move Emote Category Ids to Constatns
