--- src/modules/emote_menu/stores/emote-categories.js
@@ -1,17 +1,0 @@
-const EMOTE_CATEGORY_IDS = {
-  BETTERTTV: 'bttv',
-  FRANKERFACEZ: 'ffz',
-  TWITCH: 'twitch',
-  EMOJI_PEOPLE: 'emoji-people',
-  EMOJI_NATURE: 'emoji-nature',
-  EMOJI_FOODS: 'emoji-foods',
-  EMOJI_ACTIVITIES: 'emoji-activities',
-  EMOJI_TRAVEL: 'emoji-travel',
-  EMOJI_OBJECTS: 'emoji-objects',
-  EMOJI_SYMBOLS: 'emoji-symbols',
-  EMOJI_FLAGS: 'emoji-flags',
-  FAVORITES: 'favorites',
-  FRECENTS: 'frecents',
-};
-
-export default EMOTE_CATEGORY_IDS;
